Add files via upload
Huffman coding is a data compression algorithm that assigns shorter binary codes to more frequent symbols in the data and longer codes to less frequent symbols, reducing storage space or transmission bandwidth. It creates a binary tree structure, where each symbol is represented by a unique path from the root to a leaf node, allowing for efficient encoding and decoding of data.